requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.4.2
Django==1.11
django-bootstrap4==0.0.6
django-registration==2.4.1
djangorestframework==3.7.7
gunicorn==19.7.1
Pillow==5.0.0

psycopg2==2.7.4
python-decouple==3.1
pytz==2018.3
whitenoise==3.3.1
